# Conflicts:
#	Travely_Manager/app/src/main/java/com/hyeran/android/travely_manager/ReserveDetailFragment.kt
#	Travely_User/app/src/main/java/com/tooc/android/tooc/adapter/MypageMyReviewAdapter.kt

+++ Travely_User/app/src/main/java/com/tooc/android/tooc/model/mypage/ReviewLookupData.kt
@@ -16,6 +16,6 @@ data class ReviewLookupData(
         var regionIdx : Long,
         var reviewIdx : Int,
         var storeCall : String,
-        var storeIdx : Long,
+        var storeIdx : Int,
         var storeUrl : String
 )
